+++ packages/example/server/calendar.js
@@ -1,0 +1,9 @@
+class Calendar {
+  getNow() {
+    // return (new Date()).toLocaleTimeString();
+    // return (new Date()).toLocaleString();
+    return (new Date()).toLocaleDateString();
+  }
+}
+
+module.exports = { Calendar };