+++ packages/server/src/index.js
@@ -1,3 +1,0 @@
-export { ServerClient } from './client/server.js';
-export * from './decorators/index.js';
-export * from './metadata/index.js';